move program domain into core
Relocate program persistence, schedule structures, and runtime queue ownership under core/. Retain program.h as a forwarding compatibility header and update both native and embedded source discovery.
